Consolidate OWB and 40K import into single modal with format toggle
Two separate import buttons confused users on mobile (icon-only, no label)
and led to format mismatch errors when importing into the wrong parser.
Replaces both with a single "Import Army List" entry point that presents
a segmented format selector (Old World Builder / Warhammer 40,000) and
auto-detects the format from pasted text so users never have to choose manually.

Also consolidates the onboarding screen's two import CTAs into one.
