+++ js/owb-import.js
@@ -1,11 +1,6 @@
-// owb-import.js — Old World Builder army list text importer
+// owb-import.js — Old World Builder army list text parser
 
-import {
-  appData, createCollection, createList, createModel,
-  addModelToList, getModelType, GAME_SYSTEMS,
-  getAllFolders, createFolder
-} from './data.js';
-import { showModal, closeModal, toast } from './ui.js';
+import { GAME_SYSTEMS } from './data.js';
 
 const GAME_SYSTEM_PATTERNS = [
   { pattern: /warhammer.*old world/i,  id: 'old_world' },
@@ -113,151 +108,4 @@ export function parseOwbList(text) {
   }
 
   return { armyName: armyName || 'Imported Army', gameSystemId, units };
-}
-
-function doImport(text, folderId = null) {
-  const { armyName, gameSystemId, units } = parseOwbList(text);
-
-  if (!units.length) return null;
-
-  // Find an existing collection for this game system, or create one
-  let collectionId = Object.values(appData.collections).find(c => c.gameSystemId === gameSystemId)?.id;
-  if (!collectionId) {
-    const sys = GAME_SYSTEMS[gameSystemId];
-    collectionId = createCollection({ name: sys?.label || 'Imported', gameSystemId });
-  }
-
-  const listId = createList({ name: armyName, collectionId });
-
-  for (const unit of units) {
-    const typeObj = getModelType(unit.modelTypeId);
-    const stages = typeObj ? typeObj.stages.map(s => ({ ...s })) : null;
-    const modelId = createModel({
-      name: unit.name,
-      quantity: unit.quantity,
-      gameSystemId,
-      modelTypeId: unit.modelTypeId,
-      stages,
-      folderId,
-    });
-    addModelToList(listId, modelId);
-  }
-
-  return { collectionId, listId, unitCount: units.length, armyName };
-}
-
-const TYPE_LABELS = {
-  infantry: 'Infantry',
-  cavalry: 'Cavalry',
-  monster: 'Monster',
-  character: 'Character',
-  warmachine: 'War Machine',
-  chariot: 'Chariot',
-  swarm: 'Swarm',
-};
-
-// Show the import modal. onSuccess(collectionId, listId) is called after a successful import.
-export function showOwbImportModal(onSuccess) {
-  const folders = getAllFolders();
-
-  const content = document.createElement('div');
-  content.innerHTML = `
-    <p style="font-size:0.85em;color:var(--text-muted);margin-bottom:0.75em">
-      Paste your army list from the Old World Builder app or website. Units are imported into a new army list; characters, cavalry, and war machines are detected automatically.
-    </p>
-    <div class="form-group" style="margin-bottom:0.75em">
-      <label style="font-size:0.85em;font-weight:600">Import into folder</label>
-      <select id="owbFolderSelect" class="form-input">
-        <option value="">— Unfiled —</option>
-        ${folders.map(f => `<option value="${f.id}">${f.name}</option>`).join('')}
-        <option value="__new__">+ New folder...</option>
-      </select>
-    </div>
-    <textarea id="owbPasteArea" class="form-input" rows="12" placeholder="===
-Clan Eshin [748 pts]
-Warhammer: The Old World, Skaven...
-===
-
-++ Characters [166 pts] ++
-
-Skaven Chieftain [51 pts]
-..."></textarea>
-    <div id="owbPreview" style="margin-top:0.75em;display:none">
-      <div id="owbPreviewContent"></div>
-    </div>
-    <div class="modal-actions" style="margin-top:1em">
-      <button class="btn btn-primary" id="owbImportBtn">Import List</button>
-      <button class="btn" id="owbCancelBtn">Cancel</button>
-    </div>
-  `;
-
-  // Inline folder creation
-  content.querySelector('#owbFolderSelect').addEventListener('change', e => {
-    if (e.target.value === '__new__') {
-      const name = prompt('New folder name:');
-      if (name?.trim()) {
-        const fid = createFolder(name.trim());
-        const opt = document.createElement('option');
-        opt.value = fid;
-        opt.textContent = name.trim();
-        opt.selected = true;
-        const newOpt = e.target.querySelector('[value="__new__"]');
-        e.target.insertBefore(opt, newOpt);
-        e.target.value = fid;
-      } else {
-        e.target.value = '';
-      }
-    }
-  });
-
-  const textarea = content.querySelector('#owbPasteArea');
-  const preview = content.querySelector('#owbPreview');
-  const previewContent = content.querySelector('#owbPreviewContent');
-
-  const updatePreview = () => {
-    const text = textarea.value.trim();
-    if (!text) { preview.style.display = 'none'; return; }
-    try {
-      const { armyName, gameSystemId, units } = parseOwbList(text);
-      if (!units.length) { preview.style.display = 'none'; return; }
-      const sys = GAME_SYSTEMS[gameSystemId];
-      previewContent.innerHTML = `
-        <div style="font-size:0.82em;color:var(--text-muted);margin-bottom:0.4em">
-          Preview — <strong style="color:var(--text)">${armyName}</strong>
-          &nbsp;<span class="sys-tag ${sys?.theme || ''}">${sys?.shortLabel || gameSystemId}</span>
-          &nbsp;· ${units.length} unit${units.length !== 1 ? 's' : ''}
-        </div>
-        <div style="display:flex;flex-direction:column;gap:0.2em;max-height:180px;overflow-y:auto">
-          ${units.map(u => `
-            <div style="display:flex;gap:0.5em;font-size:0.82em;align-items:center">
-              <span style="min-width:2em;text-align:right;color:var(--text-muted)">×${u.quantity}</span>
-              <span>${u.name}</span>
-              <span style="margin-left:auto;color:var(--text-muted);font-size:0.9em">${TYPE_LABELS[u.modelTypeId] || u.modelTypeId}</span>
-            </div>
-          `).join('')}
-        </div>
-      `;
-      preview.style.display = '';
-    } catch {
-      preview.style.display = 'none';
-    }
-  };
-
-  textarea.addEventListener('input', updatePreview);
-
-  content.querySelector('#owbImportBtn').addEventListener('click', () => {
-    const text = textarea.value.trim();
-    if (!text) { toast('Please paste an army list first', 'error'); return; }
-    const rawFolder = content.querySelector('#owbFolderSelect').value;
-    const folderId = rawFolder && rawFolder !== '__new__' ? rawFolder : null;
-    const result = doImport(text, folderId);
-    if (!result) { toast('No units found — check the list format', 'error'); return; }
-    closeModal();
-    toast(`Imported "${result.armyName}" — ${result.unitCount} unit${result.unitCount !== 1 ? 's' : ''}`, 'success');
-    onSuccess(result.collectionId, result.listId);
-  });
-
-  content.querySelector('#owbCancelBtn').addEventListener('click', () => closeModal());
-
-  showModal({ title: 'Import from Old World Builder', content, wide: true });
 }
