+++ js/w40k-import.js
@@ -1,11 +1,4 @@
-// w40k-import.js — Warhammer 40,000 app army list importer
-
-import {
-  appData, createCollection, createList, createModel,
-  addModelToList, getModelType, GAME_SYSTEMS,
-  getAllFolders, createFolder
-} from './data.js';
-import { showModal, closeModal, toast } from './ui.js';
+// w40k-import.js — Warhammer 40,000 app army list parser
 
 function detectModelType(name, section) {
   const n = name.toLowerCase();
@@ -102,157 +95,3 @@ export function parseW40kList(text) {
   return { armyName: armyName || 'Imported Army', gameSystemId: 'wh40k', units: collectedUnits };
 }
 
-function doImport(text, folderId = null) {
-  const { armyName, gameSystemId, units } = parseW40kList(text);
-
-  if (!units.length) return null;
-
-  let collectionId = Object.values(appData.collections).find(c => c.gameSystemId === gameSystemId)?.id;
-  if (!collectionId) {
-    const sys = GAME_SYSTEMS[gameSystemId];
-    collectionId = createCollection({ name: sys?.label || 'Imported', gameSystemId });
-  }
-
-  const listId = createList({ name: armyName, collectionId });
-
-  for (const unit of units) {
-    const typeObj = getModelType(unit.modelTypeId);
-    const stages = typeObj ? typeObj.stages.map(s => ({ ...s })) : null;
-    const modelId = createModel({
-      name: unit.name,
-      quantity: unit.quantity,
-      gameSystemId,
-      modelTypeId: unit.modelTypeId,
-      stages,
-      folderId,
-    });
-    addModelToList(listId, modelId);
-  }
-
-  return { collectionId, listId, unitCount: units.length, armyName };
-}
-
-const TYPE_LABELS = {
-  character: 'Character',
-  infantry: 'Infantry',
-  monster: 'Monster',
-  vehicle: 'Vehicle',
-  walker: 'Walker',
-  swarm: 'Swarm',
-  cavalry: 'Cavalry',
-};
-
-// Show the import modal. onSuccess(collectionId, listId) is called after a successful import.
-export function showW40kImportModal(onSuccess) {
-  const folders = getAllFolders();
-
-  const content = document.createElement('div');
-  content.innerHTML = `
-    <p style="font-size:0.85em;color:var(--text-muted);margin-bottom:0.75em">
-      Paste your army list from the Warhammer 40,000 app. Units are imported into a new army list; characters, monsters, and vehicles are detected automatically.
-    </p>
-    <div class="form-group" style="margin-bottom:0.75em">
-      <label style="font-size:0.85em;font-weight:600">Import into folder</label>
-      <select id="w40kFolderSelect" class="form-input">
-        <option value="">— Unfiled —</option>
-        ${folders.map(f => `<option value="${f.id}">${f.name}</option>`).join('')}
-        <option value="__new__">+ New folder...</option>
-      </select>
-    </div>
-    <textarea id="w40kPasteArea" class="form-input" rows="12" placeholder="Niddos (1500 points)
-
-Tyranids
-Strike Force (2000 points)
-Invasion Fleet
-
-CHARACTERS
-
-Neurotyrant (125 points)
-  • Warlord
-  • 1x Neurotyrant claws and lashes
-
-BATTLELINE
-
-Hormagaunts (65 points)
-  • 10x Hormagaunt
-    • 10x Hormagaunt talons
-
-Exported with App Version: v1.53.0 (119), Data Version: v780"></textarea>
-    <div id="w40kPreview" style="margin-top:0.75em;display:none">
-      <div id="w40kPreviewContent"></div>
-    </div>
-    <div class="modal-actions" style="margin-top:1em">
-      <button class="btn btn-primary" id="w40kImportBtn">Import List</button>
-      <button class="btn" id="w40kCancelBtn">Cancel</button>
-    </div>
-  `;
-
-  // Inline folder creation
-  content.querySelector('#w40kFolderSelect').addEventListener('change', e => {
-    if (e.target.value === '__new__') {
-      const name = prompt('New folder name:');
-      if (name?.trim()) {
-        const fid = createFolder(name.trim());
-        const opt = document.createElement('option');
-        opt.value = fid;
-        opt.textContent = name.trim();
-        opt.selected = true;
-        const newOpt = e.target.querySelector('[value="__new__"]');
-        e.target.insertBefore(opt, newOpt);
-        e.target.value = fid;
-      } else {
-        e.target.value = '';
-      }
-    }
-  });
-
-  const textarea = content.querySelector('#w40kPasteArea');
-  const preview = content.querySelector('#w40kPreview');
-  const previewContent = content.querySelector('#w40kPreviewContent');
-
-  const updatePreview = () => {
-    const text = textarea.value.trim();
-    if (!text) { preview.style.display = 'none'; return; }
-    try {
-      const { armyName, gameSystemId, units } = parseW40kList(text);
-      if (!units.length) { preview.style.display = 'none'; return; }
-      const sys = GAME_SYSTEMS[gameSystemId];
-      previewContent.innerHTML = `
-        <div style="font-size:0.82em;color:var(--text-muted);margin-bottom:0.4em">
-          Preview — <strong style="color:var(--text)">${armyName}</strong>
-          &nbsp;<span class="sys-tag ${sys?.theme || ''}">${sys?.shortLabel || gameSystemId}</span>
-          &nbsp;· ${units.length} unit${units.length !== 1 ? 's' : ''}
-        </div>
-        <div style="display:flex;flex-direction:column;gap:0.2em;max-height:180px;overflow-y:auto">
-          ${units.map(u => `
-            <div style="display:flex;gap:0.5em;font-size:0.82em;align-items:center">
-              <span style="min-width:2em;text-align:right;color:var(--text-muted)">×${u.quantity}</span>
-              <span>${u.name}</span>
-              <span style="margin-left:auto;color:var(--text-muted);font-size:0.9em">${TYPE_LABELS[u.modelTypeId] || u.modelTypeId}</span>
-            </div>
-          `).join('')}
-        </div>
-      `;
-      preview.style.display = '';
-    } catch {
-      preview.style.display = 'none';
-    }
-  };
-
-  textarea.addEventListener('input', updatePreview);
-
-  content.querySelector('#w40kImportBtn').addEventListener('click', () => {
-    const text = textarea.value.trim();
-    if (!text) { toast('Please paste an army list first', 'error'); return; }
-    const folderId = content.querySelector('#w40kFolderSelect').value || null;
-    const result = doImport(text, folderId);
-    if (!result) { toast('No units found — check the list format', 'error'); return; }
-    closeModal();
-    toast(`Imported "${result.armyName}" — ${result.unitCount} unit${result.unitCount !== 1 ? 's' : ''}`, 'success');
-    onSuccess(result.collectionId, result.listId);
-  });
-
-  content.querySelector('#w40kCancelBtn').addEventListener('click', () => closeModal());
-
-  showModal({ title: 'Import from Warhammer 40,000 App', content, wide: true });
-}
